requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
bottle==0.12.25
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.5
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
fonttools==4.41.0
greenlet==3.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.10
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.0.2
Werkzeug==2.3.6
wikipedia==1.4.0
